Unifying the QE classes and tests in the quality_estimator*
