+++ src/translator/quality_estimator.h
@@ -2,11 +2,39 @@
 
 #include <vector>
 
-#include "definitions.h"
-#include "iquality_estimator.h"
+#include "annotation.h"
 #include "matrix.h"
+#include "response.h"
+#include "translator/history.h"
 
 namespace marian::bergamot {
+
+/// Interface for quality estimator
+class QualityEstimator {
+ public:
+  // Computes quality-scores using values from History and precomputed
+  // and stored tokenizations within Response.
+  //
+  // @param [in] histories: Histories obtained from translating a blob of source-text
+  // @param [inout] response: Partially constructed response, holding tokenization info
+  // for source and target. The quality-scores for each sentence obtained from source-text blob
+  // are written out as WordsQualityEstimate into response.
+  virtual void computeQualityScores(const Histories &histories, Response &response) const = 0;
+};
+
+/// The UnsurpervisedQE stands for Unsurpervised Quality Estimator model. It basically uses the negative log
+/// probabilities (logprobs) of the translator model as proxy for quality scores. Then, for a given word, it's quality
+/// score is computed by taking the mean of the negative logprobs of the tokens that make up it. The sentence score is
+/// the mean of all word's neg. logprob.
+class UnsupervisedQualityEstimator : public QualityEstimator {
+ public:
+  void computeQualityScores(const Histories &histories, Response &response) const override;
+
+ private:
+  Response::WordsQualityEstimate computeSentenceScores(const std::vector<float> &logProbs, const AnnotatedText &target,
+                                                       const size_t sentenceIdx) const;
+};
+
 // ASCII and Unicode text files never start with the following 64 bits
 constexpr std::size_t BINARY_QE_MODEL_MAGIC = 0x78cc336f1d54b180;
 /// The current Quality Estimator model is a Logistic Model implemented through
@@ -18,7 +46,7 @@ constexpr std::size_t BINARY_QE_MODEL_MAGIC = 0x78cc336f1d54b180;
 /// These variables are firstly initialized by parsing a file (which comes from memory),
 /// and then they are used to build a model representation
 ///
-class LogisticRegressorQualityEstimator : public IQualityEstimator {
+class LogisticRegressorQualityEstimator : public QualityEstimator {
  public:
   struct Header {
     uint64_t magic;             // BINARY_QE_MODEL_MAGIC
@@ -62,7 +90,23 @@ class LogisticRegressorQualityEstimator : public IQualityEstimator {
   Response::WordsQualityEstimate computeSentenceScores(const std::vector<float> &logProbs, const AnnotatedText &target,
                                                        const size_t sentenceIdx) const;
 
-  static Matrix extractFeatures(const std::vector<std::vector<float> > &wordLogProbs);
+  static Matrix extractFeatures(const std::vector<std::vector<float>> &wordLogProbs);
 };
+
+/// The createQualityEstimator method create a quality estimator
+/// By default, if the qualityFileMemory is empty it will use
+/// the unsupervised learning approach (UnsupervisedQualityEstimator).
+inline std::shared_ptr<QualityEstimator> createQualityEstimator(const AlignedMemory &qualityFileMemory) {
+  // If no quality file return simple model
+  if (qualityFileMemory.size() == 0) {
+    return std::make_shared<UnsupervisedQualityEstimator>();
+  }
+
+  return std::make_shared<LogisticRegressorQualityEstimator>(
+      LogisticRegressorQualityEstimator::fromAlignedMemory(qualityFileMemory));
+}
+
+std::pair<std::vector<std::vector<ByteRange>>, std::vector<std::vector<float>>> remapWordsAndLogProbs(
+    const std::vector<float> &logProbs, const AnnotatedText &target, const size_t sentenceIdx);
 
 }  // namespace marian::bergamot
