--- src/translator/quality_estimator_helper.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include <memory>
-
-#include "definitions.h"
-#include "iquality_estimator.h"
-
-namespace marian::bergamot {
-
-/// The createQualityEstimator method create a quality estimator
-/// By default, if the qualityFileMemory is empty it will use
-/// the unsupervised learning approach (UnsupervisedQualityEstimator).
-std::shared_ptr<IQualityEstimator> createQualityEstimator(const AlignedMemory& qualityFileMemory);
-}  // namespace marian::bergamot
